fix(nextls): allow extension to force a download
In the case that the auto updater in Next LS has a bug, this
patch allows the extension to force a fresh download through the
extension.

It sets a key in the extensions globalState, which is persisted across
restarts. If the key isn't there, then it forces the download.

To force a download a second time, you need to increment the key
version, which is basically setting a new key to use as a sentinel.